Interim commit.
Using mutable global hash table, to account for tables appearing "out of
order". However, there is _some_ order the parser needs to understand
and enforce, with respect to nested _arrays_ of tables -- to know which
array slot the nested table goes into, depends on what was defined
immediately before it.

So will probably need to revive _some_ of the parsing logic from the
previous commit, while preserving the order flexibility from this commit.